Moved the deprecated decorator in baselib
