+++ setup.py
@@ -78,7 +78,8 @@ setup(
         'numpy',
         'scipy',
         'shapely',
-        'psutil',
+        'psutil >= 0.4.1, <= 1.2.1',
+        'decorator',
     ],
     ext_modules=[geodetic_speedups, geoutils_speedups],
     include_dirs=include_dirs,
